Scale highres textures by precalculated scaleShift exponent
The highres textures should be created in a buffer with the size calculated
from the desired native size and the size of the highres texture. The size used
here should be 2**scaleShift as calculated by FindScaleFactor. This information
is later used to correctly clamp and mirror textures. Therefore,
(2**scaleShift)*(Width|Height)ToCreate is the size of the final texture and
(2**scaleShift)*(Width|Height)ToLoad is the size of the texture information
which are loaded from the external image. The mirror factor is already part of
the (Width|Height)ToCreate value and it must not be multiplied by it again to
get the texture buffer size.