Core update
- From main repo:

- Bring the new savestates improvements to PJ64 savestate saving.
In addition, this fixes two other bugs to the PJ64 saving code: A out-of-bounds error when writting zeroes to the FCR1..31 field, and a uninitialized value error when writting the EntryDefined field of the TLB.

- Finished the new savestate model.
By the way, last commit also changed the way Project64 TLB entries were written, not sure why. I believe they are written correctly now since I'm writting the same register values that are set in the TLBWI/TLBWR instruction. Anyway, I don't think TLB was working anyway since EntryDefined was never set (now it's always set, so it may still not work). This needs more testing agains PJ64.
The big file size increase should stop now, and maybe some cleanup will make up for it.

- Remove byte flags from SP_register. More to follow.

- Remove the rest of the memory flags, and correct a bug which last commit uncovered.
Before last commit, SP_INT ORed sp_status_reg with 0x303 (signal1, signal2, halt, broke), but only set the byte flags for signal2, broke, halt (the line for signal1 was commented). This caused a desync between the status reg and the byteflags. Next time update_SP was called, the register was rebuilt using the flags, so signal1 was unset from the status reg.
This commit does not set the signal1 bit in the sp status reg (which is consistent with the taskdone() function on the HLE RSP and the most likely previous behaviour), but this may change behaviour (for the better).

- NOTE: This BREAKS savestates for the time being. There will be no crash in the emulator, but it causes the emu to hang.
It's definitely something to do with the NEW_DYNAREC ifdefs.

Someone who knows a lot about the new dynarec should take a look and see what's wrong. I'd look, but honestly everything is so jammed together in its source file, I can't make any sense of it.

Also, if you do fix it. It would be nice if you could do the changes in such a way that the ifdefs are not needed at all.