Simplify ScancodeDialog.java and scancode_dialog.xml
Changed scancode_dialog.xml into a compound drawable. It is basically now a TextView, but with a drawable image positioned to the left of the displayed text.

This allows for removal of LinearLayout and ImageView uses.

This also improves readability slightly, since everything is now contained within one block in the XML.