Added new preference screen for mapping buttons/axes.
A more visual layout that minimizes localization requirements.

Slightly different interactions with user that permits more flexible mapping implementations.
 - Current method: user picks an N64 button/axis, and then mupen listens for input signals and maps the first input it hears
 - New method: user presses inputs, then taps a particular N64 button/axis on screen, and then mupen adds the entry to the map

The difference is subtle but important. Imagine that the user wants to map N64 button Z to positive analog trigger.  Many analog triggers produce a value of -1 at rest, +1 at fully depressed, and float values in between.  The native gamepad listener is notified as soon as the trigger changes state from -1 to -0.99.  With the current approach, mupen maps the first thing it hears and thus maps N64 button Z to negative analog trigger (i.e. trigger at rest).  Not the desire of the user.  With the new method, the user can move the trigger through its intermediate states to the desired state, and then ask for mupen to map it.  Problem solved, and equally intuitive to the user.