New universal architecture for input mapping/listening across all hardware/Android versions/IMEs.
Inputs are collected from the hardware (keyboard, gamepad, touchscreen, xperia hw) and converted to an intermediate "input code" and "strength".  "Strength" is always a float in the range 0-1 (NOT 0-80).  For digital buttons, strength only takes on the values of 0 and 1, but analog axes can range anywhere between.  "Input code" is a signed integer that extends the current notion of "key codes".  Keys, buttons, and IME-mapped analog axes are mapped to positive input codes in the range 1-255.  Native analog axes provided by Android 3.1+ are mapped to negative input codes, with the LSB encoding axis direction.

Inputs are collected by the "InputFunnel" class and its hardware-specific subclasses.  Funnels provide a universal language for abstracting away the hardware variations.  E.g. we should be able to make a funnel for the xperia touchpad and integrate it into the architecture as well.  Input funnels are agnostic to all things N64; they merely provide "input codes" and "strengths" to the subscribed listeners.

Funneled inputs are mapped to N64 controls using the "InputMapPreference" class.  This class encapsulates the entire bidirectional map between "input codes" and N64 buttons/axes.  These maps will be persistable using Android's built-in preference architecture.  Each player's controller will be associated with its own InputMapPreference instance.

Finally, the GameActivity* classes will listen to a single (hardware-dependent) InputFunnel object, and use the InputMapPreference objects to translate from the input codes and strengths to the correct player's N64 buttons/axes, passing them to the core.

Simple!