Implemented new method for determining whether to resume last session.
With the old approach, the app shouldn't resume the last game if the
user picked a new ROM in the interim.  This required some book-keeping
in the app, which was implemented with the global resumeLastSession
flag.  The new approach uses a different auto-save file for each ROM,
so that the correct savefile will always be used for the selected ROM.
This provides an additional benefit in that the user can bounce between
different ROMs without losing auto-save state for each game.

The auto-save filename is generated from a hash of the ROM's full path.