Added support for mice & unusual gamepads/joysticks.
Some older gamepads use non-standard ranges of motion when reporting
analog axes.  For example, the Logitech Wingman Rumblepad (ancient!)
maps right stick up/down to AXIS_THROTTLE and produces values in the
range [0,1] rather than the usual [-1,1].  This mod re-scales and
normalizes the axis values to [-1,1] for all MotionEvent data.

An interesting side-effect is that mice can be mapped to the N64 analog
stick as well. Mice produce AXIS_X and AXIS_Y values just like a
gamepad's left analog stick. However, values refer to pixel location on
the screen, not float values between -1 and 1. With this modification,
mice data is rescaled so that mouse at center of screen is equivalent
to left stick at center.  In game, it's kind of funny to watch Mario
chase the mouse pointer...