core: Don't use expressions with side effects in assert(..)
Expressions which have side effects should not be part of an assert. This
includes function calls to non-pure functions or store operations. Otherwise
the execution of the code depends on whether the code is compiled with NDEBUG
or not even when the assertion would never be false.

Import from hg:mupen64plus-core 488:31c032c952d2