Major revision to native touchpad <-> java interface
Xperia PLAY re-integration almost complete.

For touch events, rather than calling touchscreenBegin, touchscreenUp,
etc., the native code now calls into java through a single method,
onNativeTouch.  The java code then reconstructs a MotionEvent object
and passes it on to OnTouchEvent listeners, just like a View would. We
encapsulate all this in a new java class, NativeInputSource, which
derives from android.view.View.  This simplifies the implementation of
TouchControllers, since they already are designed to listen to a View
for touch event data.  I.e. a TouchController can either listen to a
GameSurface or a NativeInputSource for touch events.  The former is
used for normal Activity, while the latter is used for NativeActivity.

TODO: Touching buttons and pad simultaneously does not produce desired
behavior.  (I already know how to solve it, just have to do it.)