Auto-saves stored in user-specified directory.
Auto-saves are no longer lost if the user deletes the Mupen assets or
does a clean reinstall.  AutoSaves are stored in a subdirectory of the
user-specified game-save directory (like slot saves and manual saves).

Also changed the name of the auto-save file to be the name of the ROM,
rather than hash code.  That way, if the user moves the ROM folder, the
linkage to the auto-savefile isn't lost.