Revised menu, strings, much more intuitive IMO
Lioncash - I converted some of the Japanese strings to English to high-
light what needs to be re-translated.

Change summary:
- Renamed "touchscreen" settings to "touch" settings
- Moved Xperia PLAY touchpad settings into "touch" menu
- Renamed "controller" settings to "input" settings
- Revised summary text for more parallelism

It must be intuitive to users that buttons and joysticks are mapped in
the "input" section, regardless of whether the buttons are physically
attached to the phone/tablet or are part of a peripheral gamepad/key-
board. Consolidating touchpad and touchscreen controls is also more
intuitive IMO since they behave nearly identically, and some settings,
like octagonal constraints, apply to both. The main driving force is
that it should be obvious to Xperia PLAY users where buttons get mapped
and where touchpad is configured.

The default setting for touchpad is 'enabled' so that xperia play users
get out-of-box goodness. Non XPlay devices already have this preference
disabled programmatically, so this isn't a problem for them.