Added infrastructure, to filter inputs based on peripheral ID.
This is half the solution for good multi-player controls.  The other
half is to create a GUI that allows users to map a peripheral device to
a player number.  In Jelly Bean, we can permanently store this ID since
it is persistent across reboots.  For previous APIs users will need to
remap device to player every session.

This implementation does not apply to pre-Gingerbread devices, since
older APIs don't provide any sort of input identifier (as far as I can
tell in the little time I spent researching it).