frontend: Mark all overriden interface functions
We should use a consistent Java 6 style of marking all implementations
of abstract interfaces as "@Override". This is a problem for Java 5
and earlier because it cannot be used there for for anything else
but overridden functions of a superclass. Therefore the project specific
settings for the compiler in eclipse
("org.eclipse.jdt.core.compiler.source") has to be set to "1.6"