Temporary workaround for ASDP bug
This is a temporary workaround for the Activity Stack Double Pop behavior that occurs when the emulator shuts down, resulting in both the game activity and the menu activity closing.  Average user will now be unaware there is a problem.  May require tweaking the thread sleep times to avoid race condition (yes, this is a very ugly hack!)