Changed UserPrefs use pattern from static singleton to local field.
This matches the pattern used by AppData.

Wasn't enamored with the singleton-based interface for UserPrefs that I
added last night.  The pattern used by AppData seems better.  Seems to
be more direct and straightforward; scoping and side-effects are easier
to understand.

Also for some classes that read user prefs, I changed their constructor
or init classes to accept just the fields needed, rather than giving
them access to the entire prefs structure.  There are pros and cons to
this. Maintenance is a little more tedious if more pref fields are
needed later, but it also makes it very obvious up front what specific
preferences are used by a class.  If people like it better the old way,
I'm not opposed.

With these updates, it is clear now that NativeMethods and CoreInterface
can be merged into a single class, since NativeMethods is such a thin
wrapper over CoreInterface.