front: Added button mapping screens for Xperia PLAY.
These are just tweaked versions of the normal versions. The only
difference is that the c-pad and analog stick icons have been removed
and an explanation text box put in its place. This should make it clear
for XPlay owners that they do not need to map the c-pad and analog
stick when their touchpad is enabled.