front: Bugfix - Unable to map some analog inputs.
Bug reported here https://github.com/paulscode/mupen64plus-ae/issues/16
as well as numerous times on the forum.

Due to a quirk in Android, analog axes whose center-point is not zero
(e.g. an analog trigger whose rest position is -1) still produce a zero
value at rest until they have been wiggled a little bit. After that
point, their rest position is correctly recorded. The problem is that
LazyProvider calibrates the rest position of each analog channel based
on the first measurement it receives. As a workaround, we provide a
calibration button, which makes the user go through a little dance to
ensure all analog axes are pressed, then re-centers itself.

The word "Calibrate" is a bit of a misnomer here (Re-center is more
accurate) but I think users would be more inclined to hit a "calibrate"
button than a "re-center" button when things "aren't working".