front: Tweak calibration dialog mechanics, improve readability.
Previously, when the calibration dialog closed, it would also close the
button mapping screen, which is both confusing to the user and loses
the dirty map they may have been in the middle of defining.  This fixes
that by reopening the mapping screen, and restoring the dirty state of
the map. Also added a "cancel" button to the calibration screen. Though
the consequences of pressing OK when user wishes to cancel are trivial,
adding the cancel button provides provides peace of mind and a sense
that the user is always in control.

Moved some of the nested anonymous listeners to the parent class, for
readability.