Bugfix: Fixed hang at launch due to orientation change.
Doh! There's a good reason Paul put these in in the first place. I
should have just trusted him. Well, now there's an explanation I guess.

Prior to this fix, bug occurs by holding the phone in an orientation
that is different from the one selected in the video preferences, then
launching the game. The game activity launches, requests an orientation
change, and by default Android tries to restart the game activity...
leading to the stall. To circumvent this we tell Android not to restart
game activities on orientation change. Since API 13 this also means we
need to tell it not to restart for screen size changes (API 13 treats
orientation change as a size change as well).

I suppose we could tell Android to never restart game activities, by
setting all the manifest flags... but somehow that seems excessive. The
"Android way" would be to fix the restart issue altogether and permit
game activities to be restarted without issue. Don't know if that's
practical for our app, but that's google's "preferred" solution. Just
to keep in mind in case it becomes practical at some point.

References:
http://developer.android.com/guide/topics/manifest/activity-element.html