Feature: Back button mappable for pre-Honeycomb devices
This is especially important for the Xperia Play since one of its face
buttons (the circle) is a hardware Back key.

TODO: HC+ devices can't use the Back key in-game since it's used to
toggle the Action Bar, a critical element of the UI.  However an HC+
user is still able to map the Back button in the mapping screen. I
could easily add the same logic for that screen... but hoping for a
cleaner solution that doesn't involve so much special-case handling.
For example,
 - use an alternative method to opening the action bar, such as a down-
   swipe, thus freeing the back button for all devices
 - just allow users to map these buttons, and they'll figure it out on
   their own when they fail in game.
 - be more formal about what buttons can and can't be mapped, i.e. make
   a black-list class that is then shared by all key handlers.