Implemented simpler mechanism for handling special keycodes.
This largely addresses the TODO in the previous commit,
5f2c99fee69d84e8be9c996f0959dcf74a1210af

Now if a key is unmappable, either by preference (vol keys) or by GUI
requirements (back, menu keys), we get the same behavior in-game as we
do in the mapping screens.  Also a much easier approach to maintain, as
unmappable keys are all defined in one place: UserPrefs.java.

Also changed the name from isVolKeysEnabled => isVolKeysMappable for
clarity.