Enhanced PathPreference implementation (more portable, more robust)
The default values are provided in the XML file now, rather than in the
java code, making this much more modular and general. Also provides two
options for handling non-existent path (e.g. fresh install, user moved
the file/directory):
 1 - Fall-back to the external storage directory (default)
 2 - Create the parent directories needed to make the path exist

The former is useful for directories the app will be reading from
(e.g. a directory for user-provided inputs, where the user may have
moved the files between sessions).

The latter is useful for directories that the app will be writing to
(e.g. a directory for cache files, where the dev wants to ensure the
app doesn't clutter the storage root).