Implemented default values for InputMapPreference.
The map defaults can now be specified in the XML resource, rather than
the java code. As an example I created default serializations of both
the PS3 and the Xbox360 controllers.

Right now, the off-the-shelf configuration is Player 1 enabled, using a
PS3 controller. So for those users, they don't even need to go into the
mapping screen after installation, they can dive right into a game.