front: New asset extractor implementation.
This copies files from the assets folder onto the SD card.  This is not
a standard file copy; it actually involves decompressing files straight
from the APK using the AssetManager class from the Android SDK.

The advantage of this approach vs. the existing DataDownloader class is
that asset files can be stored in the git repo uncompressed, permitting
tracking and reducing the size of the local git database. Previously,
even the smallest change to the asset files required committing a
completely new (large) binary zipfile to the repository.