front: Wired special function input handling (many just stubs for now).
See the TODOs in PeripheralController for the functions that need to be
implemented.  There's probably more than one good way to implement them
so I'll leave that to you guys since you're more acquainted with the
core implementation.

One remaining problem with the special function mapping is that it's
possible to map the same key to multiple special functions, or a
special function and an N64 command. In either case, the special func
takes precedence over the N64 command and the special funcs at the
bottom take precedence over the funcs at the top of the menu.

Consider this a temporary solution, enough to get a Beta out the door.
I think the better solution ultimately would be to extend the layout of
the button mapping screen for Player 1. I'll probably tackle that once
I take care of the other more pressing tasks like multiplayer.