front: True multiplayer support added.
Inputs are now filtered by hardware ID to ensure that commands are
associated with the correct player.  This should work even with
identical controllers, though we'll have to test to be sure.

Peripheral->Player mapping is done from an in-game menu rather than the
main menu.  This menu is only visible when more than one player is
enabled in the main menu.  The in-game menu will only list the players
that are enabled.  This keeps the menu clean for single-player games.

At the moment, the app does not remember the hardware mappings between
sessions, so the user will have to reassign players to controllers
every time they restart the app.  I'll probably make this persistent
at some point to minimize the annoyance.  Regardless, if a controller
is re-plugged it will need to be re-mapped to the player, so there
will always be some annoyance factor.  This is just a limitation of
Android as far as I can tell.  JellyBean provides persistent hardware
identifiers, but they are not necessarily unique if the controllers are
the same make and model...