front: Removed calibrate button from mapping screen.
The complexity may not be worth the benefit.  It solves an issue with
Xbox controllers and perhaps some others, but adds a lot of confusing
logic to the implementation and is usually a red herring to users. The
same functionality can be achieved by restarting the app.  Plus this is
all a band-aid for the root problem, which at this point is simpler.