front: Localized and provided missing text for all mapping buttons.
For some buttons (d-pad, c-pad, analog) the text is hidden, but still
used in the title of the listener popup. This text can also be used for
the OUYA version of the mapping screen.

Adjusted arrow images in the analog pad to be more clear.

Simplified portrait version of the mapping screen to reuse resources.

Removed some orphaned strings in the Japanese translation.