front: Moved crash tester into a separate class.
BugSense classifies errors by their stack trace. If so much as a line
number changes in the stack trace, the error will be categorized as a
new error. That's undesirable for benign crash tests - we don't want a
new error report every time the lines get shuffled. So we move the
crash test machinery into a separate class that will remain line-stable
for a long time.