front: Refactored cheats menu for better load performance.
Play menu is now implemented as a separate activity: PlayMenuActivity.
This activity is distinct from, and launched by, MenuActivity.  Every
time the Play menu is opened (i.e. PlayMenuActivity is launched) it
populates the cheat options programmatically using a string map and the
ROM's CRC as a key. The string map itself is stored in a config file
that the activity can read.

Calculating the CRC can be a slow process on some devices, so when it's
finished computing it, the activity persists the result -- and the
associated ROM filename -- to permanent storage. (Persisting is
currently implemented in the AppData class, following the usual pattern
employed there.) Whenever the Play menu activity is launched, it first
compares the user's selected game to the last filename. If they match,
the activity just uses the previously computed CRC value rather than
recompute it.

Two limitations: Presently the implementation loses the user's cheat
selections when the Play menu is exited (user returns to main menu).
There is an easy solution to this and will be implemented shortly.
Second, it still takes some time to parse the config file and populate
the cheats menu items.  This will be addressed very soon by persisting
more data to further reduce the number of redundant computations.