bugfix: I think this fixes a cast exception seen in bugsense.
If a user had previously set a binary cheat, it was persisted as a
boolean in the datafile.  Now all cheats are persisted as integers,
even if they're binary.  A user with an out-of-date datafile would have
gotten a class cast exception when reading the boolean as an integer. I
*think* this fixes that problem.  Note that this only affects alpha
testers who have set binary cheats in previous release candidates.