front: Added InputMapActivity, soon to supersede InputMapPreference.
The InputMapPreference UI is too limiting for the many functions needed
during input mapping. This is because it uses a neutered version of an
AlertDialog provided by DialogPreference. We need more control, namely
a simple way to add extra menus and callbacks, without the limitations
of the AlertDialog UI and without the show/hide semantics forced by
DialogPreference.

So we give the input mapping process its own activity.  This commit
provides most of the implementation for that activity, but stops short
of actually replacing the old InputMapPreferences in the menu.  This
commit provides a possible branch point for testing the new activity
while safely maintaining all of the existing UI and functionality.