front: Replaced InputMapPreference with InputMapActivity.
Controller enabled/disabled is persisted as its own boolean now, rather
than being serialized with the input map. This is much more logical and
easier to maintain in the long run.

Fewer buttons/axes are mapped by default now, to minimize conflicts out
of the box.  They are
 - analog stick <-> AXIS_X and AXIS_Y
 - d-pad <-> KEY_DPAD_* (no AXIS_HAT_*, conflicts w/ Nyko PlayPad)
 - start, R, L buttons to their virtually universal keycodes

A TogglePreference was implemented.  Basically a fancier version of a
checkbox preference.  Besides the slider/toggle widget, the main
difference is that clicking on the list item in the preference menu
does not do anything (checkbox preference it toggles the check state).
This is so that you can handle the click in your own custom way, e.g.
to launch an activity like InputMapActivity.  To change widget state,
you need to click on the widget itself. This imitates the "master on/
off switch" pattern used in the Android settings:
 http://developer.android.com/design/patterns/settings.html