bugfix: Use appropriate Locale for machine-readable strings.
I had assumed the appropriate choice was ENGLISH, then I came across
this while looking into the Locale API.  From
http://developer.android.com/reference/java/util/Locale.html

Quote:
The default locale is not appropriate for machine-readable output. The
best choice there is usually Locale.US – this locale is guaranteed to
be available on all devices, and the fact that it has no surprising
special cases and is frequently used (especially for computer-computer
communication) means that it tends to be the most efficient choice too.