front: Added PS3 AXIS_GENERIC_n to the ignore list.
These are the pressure-sensitive face buttons of the PS3 controller.
They are rather buggy on both my official and my knock-off PS3
controllers, only giving readings for three of the four d-pad buttons
and nothing for the ABXY buttons.  So we'll just ignore them so they
don't get mapped accidentally.