front: Added in-app change log with popup dialog on update.
The change log itself is a text file in the assets folder. This file is
only extracted to memory, which is why it's not in the mupen64plus_data
folder.

The last run version is persisted and compared against the current app
version. If they differ, it pops up a changelog with just the new
changes since last run (can span multiple versions, e.g. a user going
from 2.0.2 to 2.0.4 would see the changes made in 2.0.3 and 2.0.4). The
persisted code is updated to the current code, so this is only shown
the very first run after an update.

I feel this is an important feature, as it may not be obvious to a user
when a certain game has been fixed, a certain controller has been
supported, new video settings are available, etc.  It also shows that
the project is alive and continually being improved, and that it is
always a work in progress.