front: Minimalist implementation of joystick-navigable mapping menu.
There are two ways I envision modifying the input mapping screen for
the OUYA version.
 - Tweak the existing design so that pressing the d-pad simply moves
   the focus to various buttons (rather than highlighting the buttons
   it is mapped to).  The user just focuses on the button they want to
   map and presses the select key.  Requires minimal changes to the
   code.
 - Provide an alternative layout for the mapping screen, e.g. a list
   of mappable commands rather than a bunch of buttons.  This requires
   more modification to the source but would probably provide the best
   user experience.

The second option is mostly implemented on a local git branch, but is a
little cludgy and could really use some polishing/refactoring. Since
it's so easy to do, I implemented the first solution as well, which is
represented by this commit.