front: Added a preference to manually override OUYA mode.
Preference located in the Advanced menu.

Default value is 'auto', which just determines OUYA mode based on
hardware configuration.  OUYA users can use the standard interface if
they wish, and non-OUYA users can use the OUYA interface if they wish.
The latter case is an important one, not just for debugging the OUYA
version, but also for users who want to plug their phone/tablet into a
television and use only a gamepad for interaction.