front: Disabled auto-hold for xperia Play touchpad.
If auto-hold is a desired feature for the touchpad, then it should have
its own preference (it shouldn't use the touchscreen's preference).
Since it doesn't have its own preference right now, we simply disable
the feature on touchpads to avoid user confusion. My guess is that such
a feature would be less desirable on a touchpad since there is no
visual feedback to the user. Also, in the current implementation the
feature would only affect the C-pad.