Define cache path as $UserCachePath/glidehq/ The user cache path for mupen64plus is something different compared to the user data path. On Unix systems it is stored in $HOME/.cache/mupen64plus/ and all non-cache/non-config data is stored in $HOME/.local/share/mupen64plus.
The Windows+single user centric view of the original GlideHQ made it store its
cache in the local plugins directory in a special folder called "cache". This
is not available anymore and therefore the user cache directory is used
instead. A subfolder called "glidehq" is used to avoid collisions with non
GlideHQ compatible data.