front: Implemented multi-select list preference class.
Honeycomb+ provides this natively, but I wanted backwards-compatibility
and some custom features. This gets the job done without a lot of fuss.
The dialog checkboxes revert state on rotation but properly fixing that
requires many lines of code, probably not worth the effort right now.

This class will be used to define auto-holdable touchscreen buttons.