res: Added transifex automation scripts/configs.
pull-translations.sh is just a bash script to pull translations (and a
pretty clunky one at that).  It works with cygwin though.  Feel free to
improve it.

If you're running linux you can also use the Transifex command-line
tool. The config file is already set up, at least for the menu strings.
More about the tool can be found here:
http://help.transifex.com/features/client/index.html#user-client