bugfix: Use SurfaceView instead of GLSurfaceView.
We are not using the unique features of GLSurfaceView right now, so it
doesn't make sense to extend from this more specific class.  More
importantly, some Android versions are not robust to the improper way
way we are using GLSurfaceView.  For Android 4.0.4 (and only that
version) you get a null pointer exception.  We can assign dummy objects
to workaround that bug, but why bother?  So we just revert back to
extending from SurfaceView.

The crash reports related to misuse of GLSurfaceView are linked below.
Note that these are the highest-occurring unresolved bugs to date:
https://www.bugsense.com/dashboard/project/ad73db70#error/96890002
https://www.bugsense.com/dashboard/project/ad73db70#error/102142777
https://www.bugsense.com/dashboard/project/ad73db70#error/101188558