bugfix: Emulator no longer reloads stale autosave after manual sleep.
Thanks @Gillou68310 for discovering this one and solving it in the
previous commit.  This commit just reverts some unnecessary changes and
adds some documentation.

When Mupen is in the foreground and the device is put to sleep, the
core is shutdown before onPause() is called.  This prevented Mupen
from auto-saving.  When the device was woken, a stale auto-save file
was loaded, thus losing all the user's progress.  The previous commit
fixed that bug.