front: Added support for tablets and portrait mode
Required adding two additional layouts with all buttons visible.  The
tablet layouts are doubled in scale for phones in portrait mode (because
the buttons are otherwise far too tiny).  Will need to test on a variety
of screen sizes to make sure there is no button overlap.  I believe this
new system is ready for testing.