input-android: Rumble working for player 1 via phone vibrator.
Tested on Diddy Kong racing. The pattern might need a little tuning in
the Java code, but otherwise working fine. The native code is basically
finished. The only remaining native items are to re-wire the debugger
to the extern implementation, but this should wait until the front-end
plugin has been slimmed down.

Remaining items on the Java side:
 - Create UI for specifying pak type. The button mappings do not
   currently work and should probably be removed. I favor using the in-
   game menu to switch paks, rather than a mapped button, as this would
   be much easier to implement.
 - Get rumble working on peripheral devices (gamepads). This will just
   require some Java code in the *Controller classes.  Presently, only
   the phone vibrator is used, and only for Player 1 rumbles.