front: Base portrait scale off longest side
Rather than up-scaling from the shortest side in portrait mode,
scale is now always calculated from the longest side.  This allows
the controls to be consitently sized in either orientation.  This
also resulted in the controls being a tad larger in portrait mode,
so to better fit the D-Pad in the "Both" layout, I moved the L and
Z buttons up 2%.  Will need to re-test on a range of device sizes
and resolutions.