gles2rice: Cleanup whitespace to match upstream (no functional changes)
@ecsv Sven - If this messes up your workflow for pulling changes from
upstream, please let me know and we can rewind/revert this commit.

I've been staring at this code for awhile so I thought I'd push over
some of the cleanup I did in my experimental branch.  This gets rid of
a lot of diff noise.  Should make it a little easier to spot the real
changes between up and downstream.

Note that this only changes whitespace and comment text. The only
implementation that I changed was to move some variable declarations
inside some for loops, merely to match the upstream implementation.  If
you see any other changes to the implementation, please note them right
away because they probably are an error.