gles2rice: Use SDL preprocessor symbols to target correct GL variant.
Apparently, this should auto-magically detect the appropriate OpenGL
variant to use for the platform. See <SDL_config.h>.  In our case SDL
keys off of the ANDROID symbol to determine that gles2 should be used.

We should probably use the SDL_VIDEO_OPENGL* family of symbols for
targeting code to different GL platforms.

TODO: Test with other platforms to be sure the OpenGL compiler paths
are used appropriately.