[front] - bugfix & simplification: Simplify a boolean expression in CoreInterfaceNative.java.
Previously someone could have made a file named "zip" (without an extension) and it would attempt to load it as an actual ROM. Whoever does this with files other than ROMs is obviously pretty damn daft, however better get this out of the way before someone does do that.

Made it check for a file with a length of at least 5 characters in length because the minimum length for a file with a zip extension is 5 characters (ie. "a.zip")