[gles2n64] Bugfix in Textures.cpp.
Fixes an if statement.

The way it is now, cache.hash.insert(cache.bottom->crc, NULL); gets executed regardless of the condition (the ; on the if statement terminates it).

You really don't want to set the bottom-most element's data of the texture cache to NULL unless the cached texture you're actually working with is the bottom-most texture itself.