bugfix: Fixed axis scaling for official Xbox 360 wireless controller.
This controller does not produce AXIS_HAT_X and AXIS_HAT_Y signals from
the D-pad, giving it a signature unlike all other Xbox 360 controller
clones I've tried (all wired).  Instead the D-pad signals are passed as
key events for this controller instead of motion (axis) events.

In fact, the key codes from the D-pad are 0 for all directions, making
it impossible to distinguish which direction was pressed.  Will have to
make a note of this on the forum.  Unfortunately this is an O/S level
problem that we can't address, and it's probably caused by this
controller's use of a vendor-specific (non-standard) HID profile.