acra: Add library (.so) names to ACRA report.
By far the largest single cause of crashes right now is due to link
errors with the libraries at run-time (34% of all reported crashes).

https://paulscode.iriscouch.com/acralyzer-custom/_design/acralyzer/index.html#/dashboard/mupen64plusae

This just adds the names of whatever files are found in the libraries
directory. Just want to rule out the possibility that the files weren't
installed properly.