build: Cleaned up Android makefiles for legibility.
I triple checked this and verified the build, but would appreciate a
second or third set of eyes to ensure I didn't change anything.

Wanted to make it easier to read the build flags and dependencies.
There are some linker warnings with the latest NDK related to the gnu
stl and thought this might help track it down.