Surround OpenGL-specific code with SDL_VIDEO_OPENGL.
Assert(no functional changes).

Symbol defined in the platform-specific SDL headers (SDL_config*.h).

This is a lot of ifdefs, but the point here is to facilitate trace-
ability throughout the ES 2.0 integration process. The plan is to
beautify this code later, after we verify that the OpenGL codepath is
intact and the OpenGL ES 2.0 codepath behaves as intended.