ae-bridge: Initial commit.
Starting to consolidate all paulscode-specific code into a pair of
shared libraries (exports = Java->C, imports = C->Java).  A pair is
necessary due to the dependency graph with SDL, core, and front-end.

              { SDL             }
ae-imports <- {  ^ core         } <- ae-exports
              {     ^ front-end }

The goal is to minimize the upstream diff, and to simplify upstream
synchronizations and creation/replacement of plugins/dependencies.