ae-bridge: Move main entry from SDL/main/android to ae-bridge.
The version in SDL/main is just a starting example for SDL clients.
We move our customized version into ae-bridge so that we don't have to
touch the one provided by SDL.  One less conflict when syncing with
upstream SDL.

I reverted the SDL_android_main.cpp to upstream SDL revision 4999 (on
the SDL 2.0 branch) which seems to be a pretty close match to where we
began customizing.  It's no longer compiled into our app, so it doesn't
really matter what I changed it to (if at all), but this should make
synchronizing with upstream SDL a tad smoother and the git history a
little clearer.