ae-bridge: Revert exported SDL funcs and wrap functions in ae-bridge.
Rather than modifying the names of the SDL functions to match the java
package name, we just write wrapper functions that call the original
exported functions.  Again, just minimize the diff with SDL upstream.
This also makes the transition to SDL 2 a bit cleaner.

Also, I decided not to wrap onNativeAccel, since we don't use it.