gles2n64: Save per-rom window config in separate variable.
Right now, the per-rom window sizing config is based on an assumption
that the user-provided size is always 800x480.  In the future, we want
to pass the actual display resolution via the config file.  So to
preserve the current behavior, we store the per-rom parameters as a
separate field in the structure.