gles2n64: Apply bandaid for full SDL 2.0 compliance.
In the interest of progress I'm adding a quick and dirty hack to remove
the last SDL 2.0 non-compliant piece of gles2n64.  Like to come up with
a better solution.  The offending function is SDL_SetVideoMode and the
hack solution is to copy the compatibility header from mupen core, and
modify it a tiny bit to eliminate compiler errors.